requirements.txt
openai
requests
python-dotenv
discord.py 